Move DrawingLayer to separate file, delete console.logs
